feat(settings): move master password and biometric unlock to a Security section
Vault-key state notes, the Touch ID / Windows Hello toggle and unlock
control, and change-master-password move from Settings → MCP into a new
Settings → Security section — they guard all stored secrets, not just
the MCP token. The MCP section keeps the server, token, CLI, and client
config concerns.
